requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
python-multipart==0.0.9

# AI Services
google-generativeai==0.8.3
groq==0.11.0
openai==1.54.0

# Web & HTTP
httpx==0.27.0
requests==2.31.0

# Image Processing
Pillow==11.0.0

# Environment & Configuration
python-dotenv==1.0.0

# Emergency Services (NEW)
twilio==9.0.4
sendgrid==6.11.0

# Utilities
python-dateutil==2.8.2